Add type annotations to tooltip and dropdown props
