+++ packages/elements/src/components/Popovers/Dropdown.tsx
@@ -4,18 +4,40 @@ import { Button, Div, StyleObject, mergeSelectors } from 'maker-ui'
 import { Popover, PopoverProps } from './Popover'
 
 interface DropdownProps {
+  /** ID selector for the dropdown container */
   id?: string
+  /** className selector for the dropdown container */
   className?: string
+  /** A slot that lets you supply a custom button or use a callback function that lets you
+   * construct a button with the open state.
+   * @default "Dropdown"
+   */
   button?:
     | React.ReactNode
     | ((isOpen?: boolean, attributes?: object) => React.ReactNode)
+  /** If true, the dropdown content will match the width of the button.
+   * @default false
+   */
   matchWidth?: boolean
+  /** If true, the dropdown container will trap the keyboard focus until the dropdown is closed.
+   * @default false
+   */
   trapFocus?: boolean
+  /** If true, the Dropdown will close when keyboard focus leaves the component.
+   * @default true
+   */
   closeOnBlur?: boolean
+  /** Predefined transition styles that you can use to toggle the Dropdown.
+   * @default "fade"
+   */
   transition?: PopoverProps['transition']
+  /** Responsive styles that are applied to the button if a custom button is not supplied */
   buttonCss?: StyleObject
+  /** Responsive styles that are applied to the dropdown container. */
   _css?: StyleObject
+  /** Responsive styles that are applied to the nested popover container. */
   css?: StyleObject
+  /** The contents of your dropdown component */
   children: React.ReactNode
   /** Allows you to control the dropdown from an external React.useState hook*/
   controls?: [boolean, React.Dispatch<React.SetStateAction<boolean>>]
@@ -26,7 +48,6 @@ interface DropdownProps {
  * content. It returns a customizable button and the corresponding dropdown content.
  *
  * @todo - clean up external `controls` implementation
- *
  * @todo - known issue with scale transition and height measurements
  *
  * @link https://maker-ui.com/docs/elements/popovers
@@ -37,7 +58,7 @@ export const Dropdown = ({
   matchWidth = false,
   trapFocus = false,
   closeOnBlur = true,
-  transition = 'none',
+  transition = 'fade',
   id,
   className,
   buttonCss,
