+++ packages/elements/src/components/Spinner.tsx
@@ -8,8 +8,13 @@ const defaultColors = {
 }
 
 export interface SpinnerProps extends SVGProps {
+  /** All prebuilt spinner types */
   type: 'default' | 'pulse' | 'rotate' | 'blocks' | 'scale'
+  /** A number in pixels that will determine the width and height of the spinner.
+   * Each spinner is a perfect square.
+   */
   size: number
+  /** A dictionary of colors that are applied to various parts of each Spinner SVG */
   colors: {
     primary: string
     secondary: string
@@ -23,7 +28,7 @@ interface SpinnerSVGProps extends Omit<SpinnerProps, 'type'> {}
  * The `Spinner` component gives you easy access to 5 common loading indicators for a better
  * user experience during network requests.
  *
- * @todo add useTransition to this component for smooth entry and exit
+ * @todo add react-transition-group to this component for smooth entry and exit
  * @todo refactor and remove the css repetition
  *
  * @link https://maker-ui.com/docs/elements/spinner
